Feat: 날짜 범위 데이터 변환 유틸함수 분리
